fix: deleted the composable RacksIMG and put all the code in RacksDBAPI + apply all the changes in the views
