+++ racksdb/web/ui/src/views/InfrastructureDetailsView.vue
@@ -7,7 +7,6 @@ SPDX-License-Identifier: GPL-3.0-or-later -->
 <script setup lang="ts">
 import { useHttp } from '@/plugins/http'
 import { useRacksDBAPI } from '@/composables/RacksDBAPI'
-import { useRacksDBIMGAPI } from '@/composables/RacksDBIMG'
 import { ref, onMounted, watch } from 'vue'
 import SearchBar from '@/components/SearchBar.vue'
 import InfrastructureCard from '@/components/InfrastructureCard.vue'
@@ -18,7 +17,6 @@ import { Squares2X2Icon, TableCellsIcon } from '@heroicons/vue/24/outline'
 
 const http = useHttp()
 const racksDBAPI = useRacksDBAPI(http)
-const racksDBIMG = useRacksDBIMGAPI(http)
 const infrastructures: Ref<Array<Infrastructure>> = ref([])
 const infrastructureDetails: Ref<Infrastructure | undefined> = ref()
 const cardsView = ref(true)
@@ -35,7 +33,7 @@ function toggleImageModal() {
 
 async function getInfrastructureImg() {
   try {
-    const myBlob = await racksDBIMG.infrastructureImageSvg(props.name)
+    const myBlob = await racksDBAPI.infrastructureImageSvg(props.name)
     blobURL.value = URL.createObjectURL(myBlob)
   } catch (error) {
     console.error(`Error getting ${props.name}: ` + error)
